+++ src/main/typescript/data/actor/source.ts
@@ -1,4 +1,4 @@
-import type PlayerCharacterDataSource from "./playerCharacter/source.js";
+import type CharacterDataSource from "./character/source.js";
 
 /** A union for data-sources of all Actor types */
-export type WvActorDataSource = PlayerCharacterDataSource;
+export type WvActorDataSource = CharacterDataSource;